# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsapp Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://official.whatsapp.loc/engine"
        }
    ],
    "paths": {
...
        "/users/contacts": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "fetch user data",
                "description": "fetch connected user data",
                "operationId": "contacts",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Contact Id to get specific contact",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Page Size",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns contacts object."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/checkPhone": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "check whatsapp availability",
                "description": "check if phone has whatsapp or no",
                "operationId": "checkPhone",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to get whatsapp availability",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/getUserStatus": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "fetch specific user status",
                "description": "check if phone has whatsapp or no",
                "operationId": "getUserStatus",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to get status",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/getUserPresence": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "fetch specific user presence",
                "description": "check if phone has whatsapp or no",
                "operationId": "getUserPresence",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to get presence",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/getUserPicture": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "fetch specific user profile picture",
                "description": "check if phone has whatsapp or no",
                "operationId": "getUserPicture",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to get profile picture",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/blockList": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "fetch specific user block list",
                "description": "check if phone has whatsapp or no",
                "operationId": "blockList",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/blockUser": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "block user",
                "description": "block specific user",
                "operationId": "blockUser",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to block user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/unblockUser": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "unblock user",
                "description": "unblock specific user",
                "operationId": "unblockUser",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to unblock user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/users/rejectCall": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "reject call",
                "description": "reject incoming call from specific user",
                "operationId": "rejectCall",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone to unblock user",
                        "required": true
                    },
                    {
                        "name": "callId",
                        "in": "query",
                        "description": "Call ID to reject",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
...
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "fetch all messages",
                "description": "fetch user all messages",
                "operationId": "messages",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "query",
                        "description": "Message Id to get specific message",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Page Size",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns array of messages."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "channel_id": {
                "type": "apiKey",
                "description": "Enter Your Channel ID",
                "name": "CHANNELID",
                "in": "header"
            },
            "channel_token": {
                "type": "apiKey",
                "description": "Enter Your Channel Token",
                "name": "CHANNELTOKEN",
                "in": "header"
            }
        }
    }
}